Updated spurious cmd lock detection. Also fixed a few bugs
