+++ Daily Plots/components/tlm_request.py
@@ -5,26 +5,27 @@ import urllib
 from Ska.engarchive import fetch_eng as fetch
 
 
-def data_request(user_vars,msid):
+def data_request(ts,tp,data_source,msid):
     """
     Description: Request ska_eng archive for telemetry
     Input: User defined variables, MSID
     Output: dict or JSON of data
     """
-    user_vars.ts.format = "yday"
-    user_vars.tp.format = "yday"
+    data, html = None, None
+    ts.format = "yday"
+    tp.format = "yday"
     base_url = "https://occweb.cfa.harvard.edu/maude/mrest/FLIGHT/msid.json?m="
 
-    if user_vars.data_source in ("SKA High Rate", "SKA Abreviated"):
+    if data_source in ("SKA High Rate", "SKA Abreviated"):
         fetch.data_source.set("maude")
-        if user_vars.data_source == "SKA High Rate":
+        if data_source == "SKA High Rate":
             fetch.data_source.set("maude allow_subset=False")
         else:
             fetch.data_source.set("maude allow_subset=True")
-        data = fetch.MSID(f"{msid}",user_vars.ts,user_vars.tp)
+        data = fetch.MSID(f"{msid}",ts,tp)
     else:
-        url = base_url + msid + "&ts=" +str(user_vars.ts.value) + "&tp=" + str(user_vars.tp.value)
+        url = base_url + msid + "&ts=" +str(ts.value) + "&tp=" + str(tp.value)
         response = urllib.request.urlopen(url)
         html = response.read()
 
-    return data if not user_vars.data_source in ("MAUDE Web") else json.loads(html)
+    return data if not data_source in ("MAUDE Web") else json.loads(html)
